js: DRY up body code
We were previously repeting ourselves quite a bit across Request and
Response. Now almost everything happens in BodyMixin.
