+++ js/bootstrap/response.js
@@ -1,45 +1,19 @@
 import BodyMixin from 'internal:body_mixin.js';
-import { StringReadable, isBufferish } from 'internal:common.js';
-import Headers from 'internal:headers.js';
 
 const { getPrivate } = _bindings;
 
-const rawHeadersSym = getPrivate('rawHeaders');
-const headersSym = getPrivate('headers');
 const statusSym = getPrivate('status');
 const statusTextSym = getPrivate('statusText');
-const bodySym = getPrivate('body');
-const _bodyStringSym = getPrivate('_bodyString');
+
+const bodyInit = BodyMixin.init;
 
 export default class Response {
-  // #rawHeaders; // not yet instantiated
-  // #headers; // instantiated
   // #status;
   // #statusText;
-  // #body;
-  // #_bodyString;
   constructor(body, init = {}) {
     this[statusSym] = init.status || 200;
     this[statusTextSym] = init.statusText || 'OK';
-    if (!(init.headers instanceof Headers)) {
-      this[rawHeadersSym] = init.headers;
-    } else {
-      this[headersSym] = init.headers;
-    }
-    if (body instanceof ReadableStream || body instanceof TransformStream) {
-      this[bodySym] = body;
-    } else if (typeof body === 'string') {
-      this[_bodyStringSym] = body;
-    } else if (isBufferish(body)) {
-      this[bodySym] = new StringReadable(body);
-    }
-  }
-
-  get headers() {
-    if (!this[headersSym]) {
-      this[headersSym] = new Headers(this[rawHeadersSym]);
-    }
-    return this[headersSym];
+    bodyInit.call(this, body, init);
   }
 
   get status() {
@@ -49,14 +23,5 @@ export default class Response {
   get statusText() {
     return this[statusTextSym];
   }
-
-  get body() {
-    return this[bodySym];
-  }
-
-  get _bodyString() {
-    return this[_bodyStringSym];
-  }
-
 }
 BodyMixin.mixin(Response);
